CHG: Refactor domain services to application usecases and update usecase logic
