+++ code.verse/src/main/java/app/code/verse/application/usecases/DeleteEmployee.java
@@ -1,8 +1,9 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
 import app.code.verse.domain.model.Employee;
 import app.code.verse.domain.model.utils.ValidateDataUtil;
 import app.code.verse.domain.ports.EmployeePort;
+import app.code.verse.domain.services.FindEmployee;
 
 public class DeleteEmployee {
     private EmployeePort employeePort;
