+++ code.verse/src/main/java/app/code/verse/application/usecases/DeletePolicy.java
@@ -1,7 +1,8 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
 import app.code.verse.domain.model.utils.ValidateDataUtil;
 import app.code.verse.domain.ports.PolicyPort;
+import app.code.verse.domain.services.FindPolicy;
 
 public class DeletePolicy {
     private PolicyPort policyPort;
