+++ code.verse/src/main/java/app/code/verse/application/usecases/ListEmployee.java
@@ -1,6 +1,7 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
 import app.code.verse.domain.model.Employee;
+import app.code.verse.domain.services.FindEmployee;
 
 public class ListEmployee {
     private FindEmployee findEmployee;
