+++ code.verse/src/main/java/app/code/verse/application/usecases/ListPatient.java
@@ -1,7 +1,7 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
-import app.code.verse.domain.model.Employee;
 import app.code.verse.domain.model.Patient;
+import app.code.verse.domain.services.FindPatient;
 
 public class ListPatient {
     private FindPatient findPatient;
