+++ code.verse/src/main/java/app/code/verse/application/usecases/UpdateEmployee.java
@@ -1,9 +1,10 @@
-package app.code.verse.domain.services;
+package app.code.verse.application.usecases;
 
 import app.code.verse.domain.model.Employee;
 import app.code.verse.domain.model.utils.ValidateDataUtil;
 import app.code.verse.domain.ports.EmployeePort;
 import app.code.verse.domain.ports.PersonPort;
+import app.code.verse.domain.services.FindEmployee;
 
 public class UpdateEmployee {
     private EmployeePort employeePort;
